requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
boto3==1.35.57
botocore==1.35.57
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
jmespath==1.0.1
multidict==6.1.0
mysql-connector-python==9.1.0
propcache==0.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.3
six==1.16.0
urllib3==2.2.3
yarl==1.17.1